Unit tests fixes
