+++ cloud/logger/package.json
@@ -30,22 +30,24 @@
   },
   "devDependencies": {
     "@types/bunyan": "^1.8.0",
-    "@types/mocha": "^2.2.41",
-    "@types/proxyquire": "^1.3.27",
     "@types/chai": "^4.0.0",
     "@types/loglevel": "^1.4.29",
+    "@types/mocha": "^2.2.41",
+    "@types/proxyquire": "^1.3.27",
+    "@types/sinon": "^2.3.3",
+    "chai": "^4.0.2",
     "del-cli": "^1.0.0",
     "mocha": "^3.4.2",
     "nyc": "^11.0.1",
     "proxyquire": "^1.8.0",
+    "sinon": "^2.3.5",
     "source-map-support": "^0.4.15",
     "ts-node": "^3.0.4",
+    "tslint": "^5.4.3",
     "typescript": "^2.3.4"
   },
   "dependencies": {
     "bunyan": "^1.8.10",
-    "chai": "^4.0.2",
-    "loglevel": "^1.4.1",
-    "tslint": "^5.4.3"
+    "loglevel": "^1.4.1"
   }
 }
