+++ cloud/logger/test/BunyanLogger.ts
@@ -1,26 +1,30 @@
 /// <reference types="mocha" />
 'use strict';
 
-import {BunyanLogger} from '../src/BunyanLogger';
-const loggerObject: object = {test: 'test'};
-const emptyObject: object = {};
+import * as chai from 'chai';
+import * as sinon from 'sinon';
+import { BunyanLogger } from '../src/BunyanLogger';
+const expect = chai.expect;
 
-const log = new BunyanLogger({name: 'test'});
+const loggerObject: object = { index: 'index' };
+
+const log = new BunyanLogger({ name: 'testLogger' });
 
 // call the different end points for logger
 describe('Expected results', () => {
- it('should log the expected messages', () => {
-   log.debug('debug logger message', loggerObject, loggerObject);
-   log.error('error logger message', loggerObject);
-   log.info('info logger message', loggerObject);
-   log.warn('warn logger message', loggerObject);
-   log.debug('DEBUG');
-   log.info('INFO');
-   log.warn('WARN');
-   log.error('ERROR');
+
+  it('should log the expected messages', () => {
+    log.debug('Bunyan debug logger message', loggerObject);
+    log.error('error logger message', loggerObject);
+    log.info('info logger message', loggerObject);
+    log.warn('warn logger message', loggerObject);
+    log.debug('DEBUG');
+    log.info('INFO');
+    log.warn('WARN');
+    log.error('ERROR');
   });
 
- it('should log nothing for empty messages', () => {
-   log.debug('', emptyObject);
- });
+  it('should log nothing for empty messages', () => {
+    log.debug('');
+  });
 });
